requirements.txt
flask==2.2.2
psycopg2==2.9.5
flask-sqlalchemy==3.0.2
flask-bcrypt==1.0.1
marshmallow==3.19.0
pyjwt==2.6.0
werkzeug==2.2.3
